live_scrobbler/requirements.txt
spotipy==2.23.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0 